+++ app/src/main/java/woowacourse/movie/presentation/utils/DateConverter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.utils
+package woowacourse.movie.presentation.utils
 
 import java.time.LocalDate
 import java.time.format.DateTimeFormatter
