+++ app/src/main/java/woowacourse/movie/presentation/utils/ImageConverter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.utils
+package woowacourse.movie.presentation.utils
 
 import android.content.Context
